remove dataStore in regex predicate, move logic of generating lucene query to regex predicate from regex matcher
